requirements.txt
aiofiles
aiogram
aiohttp
aiosignal
annotated-types
async-timeout
attrs
cachetools
certifi
cffi
charset-normalizer
cryptography
frozenlist
google-auth
google-auth-oauthlib
gspread
httplib2
idna
magic-filter
multidict
oauth2client
oauthlib
pip
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyparsing
python-dotenv
requests
requests-oauthlib
rsa
setuptools
six
typing_extensions
urllib3
yarl